Moved codebase out of the folder. Deleted package-lock. Added empty pages to fix render issue.
