+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\index.js":"1","D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\App.js":"2","D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\reportWebVitals.js":"3","D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\pages\\About.js":"4","D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\pages\\Main.js":"5","D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\pages\\Portfolio.js":"6","D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\pages\\Contact.js":"7"},{"size":517,"mtime":1610566654356,"results":"8","hashOfConfig":"9"},{"size":718,"mtime":1610567246317,"results":"10","hashOfConfig":"9"},{"size":375,"mtime":1610566654358,"results":"11","hashOfConfig":"9"},{"size":119,"mtime":1610567462147,"results":"12","hashOfConfig":"9"},{"size":116,"mtime":1610567438784,"results":"13","hashOfConfig":"9"},{"size":131,"mtime":1610567425018,"results":"14","hashOfConfig":"9"},{"size":125,"mtime":1610567451169,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1gl1gam",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\index.js",[],"D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\App.js",[],"D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\reportWebVitals.js",[],"D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\pages\\About.js",[],"D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\pages\\Main.js",[],"D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\pages\\Portfolio.js",[],"D:\\Repos\\class\\portfolio-react\\portfolio-react\\src\\pages\\Contact.js",[]]
