--- portfolio-react/.eslintcache
@@ -1,1 +1,0 @@
-[{"C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\index.js":"1","C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\App.js":"2","C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\pages\\Portfolio.js":"3","C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\pages\\Main.js":"4","C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\pages\\Contact.js":"5","C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\pages\\About.js":"6","C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\reportWebVitals.js":"7"},{"size":500,"mtime":1610565411772,"results":"8","hashOfConfig":"9"},{"size":584,"mtime":1610565091465,"results":"10","hashOfConfig":"9"},{"size":0,"mtime":1610564839536,"results":"11","hashOfConfig":"9"},{"size":0,"mtime":1610564806961,"results":"12","hashOfConfig":"9"},{"size":0,"mtime":1610564821742,"results":"13","hashOfConfig":"9"},{"size":0,"mtime":1610564815000,"results":"14","hashOfConfig":"9"},{"size":362,"mtime":1610565237776,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"1vktxvj",{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\index.js",[],["31","32"],"C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\App.js",[],"C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\pages\\Portfolio.js",[],"C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\pages\\Main.js",[],"C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\pages\\Contact.js",[],"C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\pages\\About.js",[],"C:\\Users\\inexp\\desktop\\bootcamp\\portfolio_react\\portfolio-react\\src\\reportWebVitals.js",[],{"ruleId":"33","replacedBy":"34"},{"ruleId":"35","replacedBy":"36"},"no-native-reassign",["37"],"no-negated-in-lhs",["38"],"no-global-assign","no-unsafe-negation"]
