+++ src/pages/About.js
@@ -1,0 +1,8 @@
+function About(){
+    return (
+        <div>
+            About
+        </div>
+        );
+}
+export default About;
