+++ src/pages/Contact.js
@@ -1,0 +1,8 @@
+function Contact(){
+    return (
+        <div>
+            Contact
+        </div>
+        );
+}
+export default Contact;
