+++ src/pages/Main.js
@@ -1,0 +1,8 @@
+function Main(){
+    return (
+        <div>
+            Main
+        </div>
+        );
+}
+export default Main;
