+++ src/pages/Portfolio.js
@@ -1,0 +1,8 @@
+function Portfolio(){
+    return (
+        <div>
+            Portfolio
+        </div>
+        );
+}
+export default Portfolio;
